python/tdd-api/pyproject.toml
[tool.poetry]
name = "tdd-api"
version = "0.1.0"
description = ""
authors = ["sarincasm <eelsweb@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
faker = "^33.3.0"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"


[tool.poetry.group.dev.dependencies]
pytest-alembic = "^0.11.1"
docker = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "tdd_api.main:start"
